backend/backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
mysqlclient
pillow
pip
psycopg2-binary
pyaes
pyasn1
PyJWT
python-dotenv
pytz
rsa
sqlparse
Telethon
tzdata